+++ prowler/CHANGELOG.md
@@ -72,6 +72,7 @@ All notable changes to the **Prowler SDK** are documented in this file.
 - Fix Validation and other errors in Azure provider [(#8915)](https://github.com/prowler-cloud/prowler/pull/8915)
 - Update documentation URLs from docs.prowler.cloud to docs.prowler.com [(#9240)](https://github.com/prowler-cloud/prowler/pull/9240)
 - Fix file name parsing for checks on Windows [(#9268)](https://github.com/prowler-cloud/prowler/pull/9268)
+- Preserve timestamps in existing findings from Security Hub integration [(#8589)](https://github.com/prowler-cloud/prowler/pull/8589)
 
 ---